[fix] PeRCNN conv3d dtype error
fix PeRCNN conv3d dtype error through mannually specify the compute dtype.
Update MindSpore package version to match CI check.

+++ MindFlow/applications/data_mechanism_fusion/percnn/gsrd_3d/train.py
@@ -139,7 +139,8 @@ def train(input_args):
                       hidden_channels=upconv_config['hidden_channel'],
                       kernel_size=upconv_config['kernel_size'],
                       stride=upconv_config['stride'],
-                      has_bias=True)
+                      has_bias=True,
+                      compute_dtype=compute_dtype)
 
     if use_ascend:
         from mindspore.amp import auto_mixed_precision
